# src/main/resources/swagger/cars.yml
openapi: "3.0.0"
info:
  title: SCars API
  description: Cars API
  version: 1.0.0

paths:
  /cars/all:
    get:
      summary: List all cars
      operationId: listCars
      tags:
        - cars
      parameters:
        -
      responses:
        '200': # status code
          description: A JSON array of user names
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CarsDTO'
  /cars/update:
    put:
      tags:
        - car-details-controller
      operationId: updateCar
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CarDTO'
  /cars/create:
    post:
      tags:
        - car-details-controller
      operationId: createCar
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CarDTO'
  /cars/{id}:
    get:
      tags:
        - car-details-controller
      operationId: getCar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CarDTO'
  /cars/delete/{id}:
    delete:
      tags:
        - car-details-controller
      operationId: deleteCar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CarDTO'
components:
  schemas:
    CarsDTO:
      properties:
        cars:
          type: array
          items:
            $ref: '#/components/schemas/CarDTO'
    CarDTO:
      type: object
      properties:
        id:
          type: integer
        carName:
          type: string
        tireSize:
          type: number
        tireType:
          type: string
          enum: [SMALL, MEDIUM, LARGE]
    ExceptionResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string